Change rate field in Currency model to decimal

+++ database/migrations/2024_11_22_121241_create_exchange_rates_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('base_currency_id')->constrained('currencies')->onDelete('cascade');
             $table->foreignId('target_currency_id')->constrained('currencies')->onDelete('cascade');
-            $table->unsignedInteger('rate')->default(0);
+            $table->decimal('rate', 12, 6)->default(0);
             $table->timestamps();
         });
     }
